Fix misc-definitions-in-headers warnings
